+++ package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@prisma/client": "^6.15.0",
+    "daisyui": "^5.1.13",
     "lucide-react": "^0.542.0",
     "next": "15.5.2",
     "react": "19.1.0",
