+++ src/app/globals.css
@@ -1,4 +1,5 @@
 @import "tailwindcss";
+/* @import "daisyui"; */
 
 :root {
   --background: #ffffff;
